Added tqdm loading bars
