+++ python/run_analysis.py
@@ -2,9 +2,9 @@ import os
 
 from urchin_utils.data_utils import dataset_by_id
 from urchin_utils.model_utils import UrchinDetector_YoloV5
-from urchin_utils.eval_utils import validiate, metrics_by_var
+from urchin_utils.eval_utils import validiate
 from urchin_utils.vis_utils import annotate_preds_on_folder
-from analysis_tools import compare_to_gt, save_detections
+from analysis_tools import compare_to_gt, save_detections, metrics_by_var
 
 
 if __name__ == "__main__":
@@ -14,15 +14,10 @@ if __name__ == "__main__":
     d = dataset_by_id("data/csvs/High_conf_clipped_dataset_V5.csv")
     yolov5_model_helio = UrchinDetector_YoloV5(weight_path)
     
-    trial_images = os.listdir("c:/Users/kelha/Downloads/images")
-    trial_images = [os.path.join("c:/Users/kelha/Downloads/images", x) for x in trial_images]
-     
-    save_detections(yolov5_model_helio, trial_images, "c:/Users/kelha/Downloads/Marlborough_Sounds_detections.csv")
-    
     
     #compare_to_gt(yolov5_model_helio, val_txt, d, "all", display_correct=True)
 
-    #validiate(yolov5_model_helio, val_txt, d)
+    validiate(yolov5_model_helio, val_txt, d)
 
     #compare_to_gt(yolov5_model_helio, val_txt, d, "all", filter_var="source",
     #              filter_func=lambda x: x == "UoA Sea Urchin")
