+++ python/urchin_utils/eval_utils.py
@@ -1,6 +1,7 @@
 import torch
 import numpy as np
 import pandas as pd
+from tqdm import tqdm   
 
 from urchin_utils.model_utils import LABEL_TO_NUM, NUM_TO_LABEL, project_sys_path, gt_to_detection
 from urchin_utils.data_utils import id_from_im_name, process_images_input
@@ -110,7 +111,7 @@ def get_metrics(model, image_set, dataset, min_iou_val = 0.5, max_iou_val = 0.95
     stats = [] #(num correct, confidence, predicated classes, target classes)
 
     #get the relavent stats for each images predictions
-    for im_path in image_set:
+    for im_path in tqdm(image_set, desc="Computing metrics", bar_format="{l_bar}{bar:30}{r_bar}"):
         #get gt and predicted boxes
         id = id_from_im_name(im_path)
         gt = gt_to_detection(dataset[id])
@@ -124,9 +125,9 @@ def get_metrics(model, image_set, dataset, min_iou_val = 0.5, max_iou_val = 0.95
             instance_counts[-1] += 1
         else:
             for box in gt:#add number of boxes of each species
-                instance_counts[box[5]] += 1
+                instance_counts[int(box[5])] += 1
 
-        target_classes = [box[5] for box in gt]
+        target_classes = [int(box[5]) for box in gt]
         correct = torch.zeros(num_of_preds, num_iou_vals, dtype=torch.bool, device=device)
 
         if num_of_preds == 0:
@@ -137,7 +138,7 @@ def get_metrics(model, image_set, dataset, min_iou_val = 0.5, max_iou_val = 0.95
         #determine which predictions are correct at each iou val
         if num_of_gts:            
             for i in range(len(iou_vals)):
-                correct[:, i] = correct_predictions(gt, pred, iou_vals[i])
+                correct[:, i] = torch.from_numpy(correct_predictions(gt, pred, iou_vals[i].item())[0]).to(device)
             
         xyxy_preds = torch.from_numpy(np.vstack([box for box in pred.gen(box_format="xyxycl")]))
         stats.append((correct, xyxy_preds[:, 4], xyxy_preds[:, 5], torch.tensor(target_classes, device=device)))
@@ -187,41 +188,4 @@ def validiate(model, images, dataset):
     print_metrics(*metrics)
     
     
-def metrics_by_var(model, images, dataset, var_name, var_func = None):
-    """Seperate the given dataset by the chosen variable and print the metrics of each partition
-    Args:
-        model: model to run
-        images: txt file of image paths or list or image paths
-        var_name: csv header name to filter by
-        var_func: optional func to run the value of var_name through, useful for discretization"""
-    
-    #read image paths from txt file
-    image_paths = process_images_input(images)
 
-    #split data by var_name
-    splits = {}
-    for image_path in image_paths:
-        #read value from dataset and apply func is given
-        id = id_from_im_name(image_path)
-        value = dataset[id][var_name]
-        if var_func: value = var_func(value)
-
-        #add image path to dict of splits based on the value
-        if value in splits:
-            splits[value].append(image_path)
-        else:
-            splits[value] = [image_path]
-
-    #print header
-    print("----------------------------------------------------")
-    print(f"Getting metrics by {var_name}{' and ' + var_func.__name__ if var_func else ''}")
-    print(f"Values: {', '.join([str(k) for k in sorted(splits.keys())])}")
-    print("----------------------------------------------------")
-
-    #print metrics for each split
-    for value in sorted(splits):
-        print(f"Metrics for {value} ({len(splits[value])} images):\n")
-        validiate(model, splits[value], dataset)
-        print("----------------------------------------------------")
-        
-    print("FINISHED")
